research/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
bleach==6.3.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
PyJWT==2.11.0
python-slugify==8.0.4
requests==2.32.5
SQLAlchemy==2.0.46
stripe==14.3.0
text-unidecode==1.3
twilio==9.10.1
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.5
yarl==1.22.0
